Cmake updates (#64)
* update some cmake code

* fix missing quote

* more some cmake files around and fix the windows issues with add suitesparse

* remove fncs library as fncs is deprecated

* update some compiler flag options

* Automated formatting of doc/script files (#65)


* move around some files for clang-tidy

* add executable and more cpplint fixes

* fix missing pragma

* Automated formatting of doc/script files

+++ src/griddyn/solvers/sundialsInterface.cpp
@@ -10,6 +10,8 @@
  * LLNS Copyright End
  */
 
+#include "sundialsInterface.h"
+
 #include "griddyn/griddyn-config.h"
 #include "idaInterface.h"
 #include "kinsolInterface.h"
@@ -391,7 +393,7 @@ namespace solvers {
             sd->nnz = a1->size();
             if (sd->flags[fileCapture_flag]) {
                 if (!sd->jacFile.empty()) {
-                    auto val = static_cast<long int>(sd->get("nliterations"));
+                    auto val = static_cast<std::uint32_t>(sd->get("nliterations"));
                     writeArray(time, 1, val, sd->mode.offsetIndex, *a1, sd->jacFile);
                 }
             }
